Revert "using Integer data type for nullable columns in Event to indicate null as no value instead of zero"
This reverts commit 018cf832efd71321b9af888a08bbb26bc694a0cd.

+++ opennms-webapp/src/main/java/org/opennms/web/event/Event.java
@@ -482,8 +482,8 @@ public class Event {
      *
      * @return a {@link java.lang.Integer} object.
      */
-    public Integer getNodeId() {
-        return (nodeID);
+    public int getNodeId() {
+        return (nodeID.intValue());
     }
 
     /**
@@ -500,8 +500,8 @@ public class Event {
      *
      * @return a {@link java.lang.Integer} object.
      */
-    public Integer getServiceId() {
-        return (serviceID);
+    public int getServiceId() {
+        return (serviceID.intValue());
     }
 
     /**
